Notes progress

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/services/CredentialServices.java
@@ -5,6 +5,7 @@ import com.udacity.jwdnd.course1.cloudstorage.model.Credential;
 import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import org.springframework.stereotype.Service;
 
+import javax.validation.constraints.NotNull;
 import java.security.SecureRandom;
 import java.util.Base64;
 import java.util.List;
@@ -20,6 +21,10 @@ public class CredentialServices {
         this.encryptionService = encryptionService;
     }
 
+    public Credential getCredential(int credentialId) {
+        return credentialMapper.getCredential(credentialId);
+    }
+
     public int addCredential(Credential credential) {
         SecureRandom random = new SecureRandom();
         byte[] salt = new byte[16];
@@ -31,18 +36,34 @@ public class CredentialServices {
     }
 
     public String getCredentialPlainTextPassword(Credential credential) {
-        return encryptionService.encryptValue(credential.getPassword(), credential.getKey());
+        return encryptionService.encryptValue(credential.getPassword(), credential.getSalt());
     }
 
     public List<Credential> getUserCredentials(User user) {
-        return credentialMapper.getUserCredentials(user);
+        List<Credential> credentials = credentialMapper.getUserCredentials(user);
+
+        for (Credential credential : credentials) {
+            credential.setPassword(encryptionService.decryptValue(credential.getPassword(), credential.getSalt()));
+        }
+
+        return credentials;
     }
 
     public void updateCredential(Credential credential) {
-        credentialMapper.updateCredential(credential);
+        Credential dbCredential = credentialMapper.getCredential(credential.getCredentialId());
+        String hashedPassword = encryptionService.encryptValue(credential.getPassword(), dbCredential.getSalt());
+        dbCredential.setUrl(credential.getUrl());
+        dbCredential.setUsername(credential.getUsername());
+        dbCredential.setPassword(hashedPassword);
+        credentialMapper.updateCredential(dbCredential);
     }
 
-    public void deleteCredential(Credential credential) {
-        credentialMapper.deleteCredentials(credential);
+    public boolean deleteCredential(@NotNull Credential credential) {
+        if (credentialMapper.getCredential(credential.getCredentialId()) != null) {
+            credentialMapper.deleteCredentials(credential);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
