+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/services/NoteServices.java
@@ -5,6 +5,7 @@ import com.udacity.jwdnd.course1.cloudstorage.model.Note;
 import com.udacity.jwdnd.course1.cloudstorage.model.User;
 import org.springframework.stereotype.Service;
 
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @Service
@@ -16,19 +17,28 @@ public class NoteServices {
         this.noteMapper = noteMapper;
     }
 
-    public int addNote(Note note) {
+    public Note getNote(int noteId) {
+        return noteMapper.getNote(noteId);
+    }
+
+    public int addNote(@NotNull Note note) {
         return noteMapper.addNote(note);
     }
 
-    public void deleteNote(Note note) {
-        noteMapper.deleteNote(note);
+    public boolean deleteNote(@NotNull Note note) {
+        if (noteMapper.getNote(note.getNoteid()) != null) {
+            noteMapper.deleteNote(note);
+            return true;
+        } else {
+            return false;
+        }
     }
 
-    public void updateNote(Note note) {
+    public void updateNote(@NotNull Note note) {
         noteMapper.updateNote(note);
     }
 
-    public List<Note> getUserNotes(User user) {
+    public List<Note> getUserNotes(@NotNull User user) {
         return noteMapper.getUserNotes(user);
     }
 }
